Fix issue #12, and another tokenization bug covered by a new test case.
The first bug was that when the maximum length of the parse buffer was set to 0,
@read (and @tokenise) would still write a result for the first tokenized word.

The second bug was that @tokenise with "flag" set to true would completely
ignore unrecognized words, instead of skipping their result blocks. That is,
"apple foo banana" (where "foo" is unrecognized) should write a 3-word result,
filling in the records for words 1 and 3 and skipping word 2, but instead it
incorrectly wrote a 2-word result, writing 'banana' as word 2.